+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+> **Note:** Dates are formatted as `yyyy-mm-dd`.
+
+## [0.0.23] - 2025-03-03
+
+- Better error handling [EXAMPLE](./examples/error.ts);
+- if `follow` property value is **true**, then it will follow redirects up to 10 times.
+- `ResponseWrapper.url` value is now final destination url even after redirects.
+
 ## [0.0.21] & [0.0.22] - 2025-02-28
 
 - Added a `local` cache mode! However redis is the default. [EXAMPLE](./examples/cache.ts)